feat(frontend): Add pagination and alert components to the frontend

+++ components/ui/Select/Select.tsx
@@ -1,4 +1,4 @@
-import React, { useState, Fragment } from "react";
+import React, { Fragment } from "react";
 import { Listbox, Transition } from "@headlessui/react";
 import {
   CheckIcon,
@@ -6,12 +6,13 @@ import {
   BarsArrowDownIcon,
   BarsArrowUpIcon
 } from "@heroicons/react/20/solid";
-import cn from "classnames";
+import { Option } from "../../../types";
+
 
 interface SelectProps {
-  options: string[];
-  onChange: React.Dispatch<React.SetStateAction<string>>;
-  value: string;
+  options: Array<Option>;
+  onChange: React.Dispatch<React.SetStateAction<Option>>;
+  value: Option;
   onSort: () => void;
   isAsc: boolean;
 }
@@ -27,7 +28,7 @@ const Select = ({ options, onChange, value, isAsc, onSort }: SelectProps) => {
               {isAsc ? <BarsArrowUpIcon className="w-5 h-5" />: <BarsArrowDownIcon className="w-5 h-5" />}
             </span>
             <Listbox.Button className="relative truncate cursor-default h-full flex-1 py-2 pr-8 rounded-lg text-left focus:outline-none sm:text-sm">
-              <span className="block truncate">{value}</span>
+              <span className="block truncate">{value.text}</span>
               <span className="pointer-events-none absolute inset-y-0 right-0 flex items-center pr-2">
                 <ChevronDownIcon
                   className="h-5 w-5 text-gray-400"
@@ -43,34 +44,30 @@ const Select = ({ options, onChange, value, isAsc, onSort }: SelectProps) => {
             leaveTo="opacity-0"
           >
             <Listbox.Options className="absolute mt-1 max-h-60 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-              {options.map((option) => (
-                <Listbox.Option
-                  key={option}
+              {options.map((option) => {
+                if(option.name === 'default') {
+                  return null;
+                }
+                return (<Listbox.Option
+                  key={option.name}
                   className={({ active }) =>
-                    `relative cursor-default select-none py-2 pl-10 pr-4 ${
-                      active ? "bg-amber-100 text-primary" : "text-gray-900"
+                    `relative cursor-default select-none py-2 px-2 ${
+                      active ? "bg-sky-100" : ""
                     }`
                   }
                   value={option}
                 >
                   {({ selected }) => (
-                    <>
                       <span
                         className={`block truncate ${
-                          selected ? "font-medium" : "font-normal"
+                          selected ? "font-medium text-gray-900" : "font-normal text-gray-700"
                         }`}
                       >
-                        {option}
+                        {option.text}
                       </span>
-                      {selected ? (
-                        <span className="absolute inset-y-0 left-0 flex items-center pl-3 text-amber-600">
-                          <CheckIcon className="h-5 w-5" aria-hidden="true" />
-                        </span>
-                      ) : null}
-                    </>
                   )}
-                </Listbox.Option>
-              ))}
+                </Listbox.Option>)
+})}
             </Listbox.Options>
           </Transition>
         </div>
